fix: return Int number in compare method (NameCompare class) so an automatic parse is not needed to be done

+++ src/main/java/second_level/exercise1And2/NameCompare.java
@@ -9,7 +9,7 @@ public class NameCompare implements Comparator<Restaurant> {
         if (!(rest1.getName().equals(rest2.getName()))){
            return rest1.getName().compareTo(rest2.getName());
         }
-        return Double.compare(rest2.getPunctuation(), rest1.getPunctuation());
+        return Integer.compare(rest2.getPunctuation(), rest1.getPunctuation());
     }
 
 }
